# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/subaccounts/{uuid}": {
            "get": {
                "tags": [
                    "Subaccounts"
                ],
                "summary": "Get a specific subaccount",
                "description": "Get a specific subaccount",
                "operationId": "ad73511451a322cf87e9d5a2dd5d43ef",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Subaccount UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subaccount retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subaccount_id": {
                                            "description": "Subaccount UUID",
                                            "type": "string",
                                            "example": "123456"
                                        },
                                        "external_id": {
                                            "description": "Subaccount ExternalId",
                                            "type": "string",
                                            "example": "123456"
                                        },
                                        "description": {
                                            "description": "Subaccount Description",
                                            "type": "string",
                                            "example": "My subaccount"
                                        },
                                        "wallet": {
                                            "properties": {
                                                "wallet_id": {
                                                    "description": "Wallet UUID",
                                                    "type": "string",
                                                    "example": "123456"
                                                },
                                                "balance": {
                                                    "description": "Wallet Balance",
                                                    "type": "string",
                                                    "example": "0.00"
                                                },
                                                "clabe": {
                                                    "description": "Wallet CLABE",
                                                    "type": "string",
                                                    "example": "123456"
                                                },
                                                "last_movements": {
                                                    "description": "Last movements",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "movement_id": {
                                                                "description": "Movement UUID",
                                                                "type": "string",
                                                                "example": "123456"
                                                            },
                                                            "card": {
                                                                "properties": {
                                                                    "card_id": {
                                                                        "description": "Card UUID",
                                                                        "type": "string",
                                                                        "example": "123456"
                                                                    },
                                                                    "masked_pan": {
                                                                        "description": "Card Masked PAN",
                                                                        "type": "string",
                                                                        "example": "123456"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "type": {
                                                                "description": "Movement Type",
                                                                "type": "string",
                                                                "example": "deposit"
                                                            },
                                                            "description": {
                                                                "description": "Movement Description",
                                                                "type": "string",
                                                                "example": "Deposit"
                                                            },
                                                            "amount": {
                                                                "description": "Movement Amount",
                                                                "type": "string",
                                                                "example": "100.00"
                                                            },
                                                            "balance": {
                                                                "description": "Movement Balance",
                                                                "type": "string",
                                                                "example": "100.00"
                                                            },
                                                            "date": {
                                                                "description": "Movement Date / Unix Timestamp",
                                                                "type": "string",
                                                                "example": "1716611739"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}